third_party/blink/web_tests/fast/loader/stateobjects/replacestate-in-iframe-expected.txt

ALERT: onpageshow
ALERT: Navigating back...
ALERT: onpageshow
PASS
